+++ changelog.md
@@ -5,6 +5,8 @@ version 5.8.0 - upcoming
   * Oauth: Authentication Grant flow now supports choosing port for return url
   * extended information on http responses (ssl certificate, body size etc)
   * support for client certificate (for testing mTls)
+  * new [Plugin management](/milkman-plugins-management) plugin - many thanks to @lopitz
+    * checkout [new themes](https://github.com/lopitz/milkman-themes) from @lopitz as well
 
 version 5.7.0
   * minor UI fixes